Move remaining endpoints into FakeIdCore
Seventh step of the Javalin decoupling. The last four endpoints
(/discovery, /jwks, /userinfo, /introspect) are now served from
framework-neutral methods on FakeIdCore. The /introspect flow uses
the IntrospectRequest and IntrospectResponse DTOs introduced in
step 2; with NON_NULL serialisation, a 'not-active' response
collapses to exactly {"active":false}, matching the prior
Map.of("active", false) wire format.

FakeIdProvider now owns a single field — the FakeIdCore — and its
endpoint methods are one-liner adapters. Construction of the oidc4j
Provider, the TokenMinter, and the JWK/keystore wiring all move
into FakeIdCore, so the adapter no longer needs to know anything
about the underlying OIDC plumbing. The only HTTP-specific logic
left in the adapter is form/query extraction, Basic-auth parsing
(ClientCredentials), and the two early-validation paths
(invalid_request, invalid_client) that gate the token endpoint.

FakeIdCore.discovery() returns DiscoveryDocument directly — the
oidc4j type is Jackson-serialisable and re-wrapping it in a Map
would force a needless conversion.

Protected by LibraryTests (discovery endpoint assertions),
UserInfoTests (userinfo JSON shape), ProviderTests, plus the
existing /token and /authorize suites. 26/26 tests pass.
